TestSulWowza: refactor validation tests into separate file
